Add string utility function formatSwissIbanNumber
