+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `formatSwissIbanNumber` string utility function
+
 ## [2.1.0] - 2025-09-03
 
 ### Added
