+++ src/lib/string.spec.ts
@@ -1,12 +1,4 @@
-import {
-  isNullOrEmpty,
-  isNullOrWhitespace,
-  capitalize,
-  uncapitalize,
-  truncate,
-  isValidSwissIbanNumber,
-  isValidSwissSocialSecurityNumber,
-} from "./string";
+import { isNullOrEmpty, isNullOrWhitespace, capitalize, uncapitalize, truncate } from "./string";
 
 describe("string tests", () => {
   test.each([
@@ -127,34 +119,5 @@ describe("string tests", () => {
     ["short", 10, "short"],
   ])("truncate without suffix parameter", (value, maxLength, expected) => {
     expect(truncate(value, maxLength)).toBe(expected);
-  });
-
-  test.each([
-    [null as unknown as string, false],
-    [undefined as unknown as string, false],
-    ["CH9300762011623852957", true],
-    ["CH93 0076 2011 6238 5295 7", true],
-    ["CH930076 20116238 5295 7", false],
-    ["CH93-0076-2011-6238-5295-7", false],
-    ["CH93 0000 0000 0000 0000 1", false],
-    ["ch93 0076 2011 6238 5295 7", false],
-    ["DE93 0076 2011 6238 5295 7", false],
-  ])("check if this swiss IBAN is valid or not", (unformattedIbanNumber, expected) => {
-    expect(isValidSwissIbanNumber(unformattedIbanNumber)).toBe(expected);
-  });
-
-  test.each([
-    [null as unknown as string, false],
-    [undefined as unknown as string, false],
-    ["7561234567891", false],
-    ["7569217076985", true],
-    ["756.92170769.85", false],
-    ["756.9217.0769.85", true],
-    ["756..9217.0769.85", false],
-    ["756.1234.5678.91", false],
-    ["test756.9217.0769.85", false],
-    ["7.56..9217...0769.85", false],
-  ])("check if the social insurance number is valid or not", (ahvNumber, expected) => {
-    expect(isValidSwissSocialSecurityNumber(ahvNumber)).toBe(expected);
   });
 });
